w5c2_ebook process
