+++ w5c2_ebook/src/slow.cpp
@@ -72,34 +72,3 @@ private:
     swap(user_positions_[lhs_id], user_positions_[rhs_id]);
   }
 };
-
-
-int main() {
-  // Для ускорения чтения данных отключается синхронизация
-  // cin и cout с stdio,
-  // а также выполняется отвязка cin от cout
-  ios::sync_with_stdio(false);
-  cin.tie(nullptr);
-
-  ReadingManager manager;
-
-  int query_count;
-  cin >> query_count;
-
-  for (int query_id = 0; query_id < query_count; ++query_id) {
-    string query_type;
-    cin >> query_type;
-    int user_id;
-    cin >> user_id;
-
-    if (query_type == "READ") {
-      int page_count;
-      cin >> page_count;
-      manager.Read(user_id, page_count);
-    } else if (query_type == "CHEER") {
-      cout << setprecision(6) << manager.Cheer(user_id) << "\n";
-    }
-  }
-
-  return 0;
-}
